Deal with partially encrypted volumes
Fix usage when the encryption of a volume is paused in the middle.
This commit also moves preparation function to another file as there's
more and more of them.

Note that Windows 8's partially encrypted volumes (as in the feature)
were already well dealt with as it displays the volume as fully
encrypted in the BitLocker's metadata header.