Better handling of virtualized regions
What I call virtualized region here is a BitLocker area reported as
filled with zeroes when looked at from the NTFS layer.

A new virtualized region appears in W$ 8 encrypted volumes. I'm not sure
what it represents, but it's here. Therefore, change the way we're
dealing with these areas to improve it, hopefully.