Add check for instable BitLocker volume
Now that it's almost sure the fields in the BitLocker metadata header
really are states, this commit adds a check for instable volumes that
dislocker might corrupt if using fuse.
Moreover, warn the user if the encryption/decryption has been paused for
her/him not to resume the state while using dislocker.