Rationalize checking the volume's state
We used to check for the BitLocker volume state only when receiving a
decryption request and printed a warning/did checks each time.
We now check the volume state upon receiving an encryption or decryption
request, and the volume state check is done only once and remembered for
the following checks.