Fix decryption while volume's state is DECRYPTED
When the volume is about to get encrypted, the volume's state is
DECRYPTED, keys don't exist, so we can get a faster path. However,
there are some things we need, such as the volume's size or volume's
stability state.
The path we took in dislocker now have these things we need.