Make bindings compile on Ubuntu 14.04
Ubuntu 12.04 doesn't seem to provide a ruby*.pc file in the ruby-dev
package. I don't have a reliable mean to find this, so we'll not support
Ubuntu 12.04 right now for the Ruby bindings. Another reason not to
support this platform is the 1.8 version of Ruby - which is known to
break API regularly.