Clarify which state we're dealing in config.h
The `dont_check_state' field may be confusing for people looking at the
code as the dis_state_e is only for the dis_initialize() function and
has nothing to do with the former.
The field has been renamed `dont_check_volume_state' to show we're
speaking of the volume and not the initialization.