Make the dislocker brew formula SSL aware
The buying of PolarSSL by ARM have made a library-name changement from
libpolarssl to libmbedtls. While I prefer up-to-date systems worldwide,
I can't expect people to update their brew package before installing
dislocker. So, here's a quite dirty way to find if libpolarssl.a is
instaled or if it's libmbedtls.a on OSX.
This also should fix travais-ci builds.