Add NULL volume encryption state
When the disk is in the decrypted state - that's when no data has been
encrypted yet, the windows user is in the menu to get the disk encrypted
but hasn't click on the "start encryption" yet - the cipher used to
encrypt the disk isn't chosen or marked in the volume metadata. So we
add the NULL state so that UNKNOWN isn't written for we know it's just
not encrypted.