Adapt FindPolarSSL.cmake to mbedTLS v2.0.0
On version 2.0.0 of mbedTLS, functions for AES or SHA have been put in a
different library than libmbedtls, namely, libmbedcrypto. However, a
typo in the CMakeLists.txt of the mbed archive prevents this
libmbedcrypto to be installed like the other libraries. Some packages
may not have shipped libmbedcrypto, such as the Homebrew PolarSSL
package.
We thus warn the user for this problem, next step is writing the
documentation on how to bypass this.