Fix VERSION_DBG in out-of-tree builds
The VERSION_DBG preprocessor token is used in order to try and reproduce
potential errors during execution by taking the git branch and commit.
It is then displayed in the firsts lines of dislocker.
When doing an out-of-tree build, the git commands to take the branch and
the commit are run from the build directory, not the source one, which
then results in an empty VERSION_DBG token, which is then rendered
useless.
This commit fixes this behavior by setting the WORKING_DIRECTORY of the
project's source during the execution of both of these git commands.