Change PATH during travis builds
During the execution of dislocker-find, the loader will call the first
ruby binary in the PATH, which can have a different ruby version against
which dislocker has been build with.
The problem is due to rvm being installed in the ubuntu VM of travis-ci,
but not fully installed.
Thus this commit to try and fix the path from which ruby will be called.