Install osxfuse through a cask
It doesn't seem possible anymore to install osxfuse through a brew
formula, we have to pass through another step. This step is a cask,
whatever that means in the homebrew world.
It doesn't seem possible to currently make a formula depends on a
cask[1], thus the added command in the INSTALL.md.

[1] https://github.com/Homebrew/legacy-homebrew/issues/49778