cmake: allow explicitly disabling Ruby bindings
Users may wish to build Dislocker without Ruby even if it's installed on
the system. This is useful for packagers, or users who don't want to
recompile Dislocker every time their package manager updates Ruby.

Add a new WITH_RUBY cmake option that can be set to ON, OFF, or AUTO.
AUTO (the default) behaves as Dislocker does currently - check for Ruby
and use it if found, ON requires Ruby and fails if not found, and OFF
never checks for Ruby and disables the bindings.