feat: support the latest FUSE on macOS
This drops osxfuse support in favor of macFUSE. macFUSE is a newer
version of osxfuse that supports the latest release of macOS, and is a
rebranded version of the same project.

This also makes the build script locate libfuse more reliably by using
pkg-config instead of searching in predetermined directories.